+++ spec/rspec/core/configuration_spec.rb
@@ -1618,5 +1618,40 @@ module RSpec::Core
       end
     end
 
+    describe "#expose_current_running_example_as" do
+      before { stub_const(Configuration::ExposeCurrentExample.name, Module.new) }
+
+      it 'exposes the current example via the named method' do
+        RSpec.configuration.expose_current_running_example_as :the_example
+        RSpec.configuration.expose_current_running_example_as :another_example_helper
+
+        value_1 = value_2 = nil
+
+        ExampleGroup.describe "Group" do
+          it "works" do
+            value_1 = the_example
+            value_2 = another_example_helper
+          end
+        end.run
+
+        expect(value_1).to be_an(RSpec::Core::Example)
+        expect(value_1.description).to eq("works")
+        expect(value_2).to be(value_1)
+      end
+
+      [:example, :running_example].each do |name|
+        it "silences the deprecation warning for ##{name} when configured to expose via that name" do
+          RSpec.configuration.expose_current_running_example_as name
+          allow(RSpec).to receive(:deprecate)
+
+          ExampleGroup.describe "Group" do
+            specify { __send__(name) }
+          end.run
+
+          expect(RSpec).to_not have_received(:deprecate)
+        end
+      end
+    end
+
   end
 end