+++ Changelog.md
@@ -12,10 +12,10 @@ Enhancements
   use during the upgrade process when external gems use the
   deprecated `RSpec::Core::ExampleGroup#example` and
   `RSpec::Core::ExampleGroup#running_example` methods. (Myron Marston)
-
-Bug fixes
-
 * Limit spamminess of deprecation messages. (Bradley Schaefer, Loren Segal)
+* Add `config.raise_errors_for_deprecations!` option, which turns
+  deprecations warnings into errors to surface the full backtrace
+  of the call site. (Myron Marston)
 
 Deprecations